#37 [RENAME] Modal 이름 SideModal로 변경
