+++ src/components/SideModal/SideModal.styles.ts
